python/requirements.txt
slackclient
slack_bolt
openai
python-dotenv
requests